Moved parse_url tests
